requirements.txt
functions-framework==3.*
google-auth==2.*
google-auth-oauthlib==1.*
google-api-python-client==2.*